backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
networkx==3.2.1
scikit-learn==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
lime==0.2.0.1
shap==0.44.0
transformers==4.36.0
torch==2.1.1
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
